Move to directory and rename fixture modules
